services/alert-manager/requirements.txt
flask==3.0.0
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
pytest-mock==3.12.0